@@ -122,12 +122,6 @@ func getPlaceholderCheck(metricName string, metricType string) *Check {
 	return check
 }
 
-type testpairCheck struct {
-	Name          string
-	resultLast    []map[string]string
-	resultCurrent []map[string]string
-}
-
 var testCheck = []Check{
 	{"test_check_a",
 		"check_a.sh",
@@ -136,8 +130,8 @@ var testCheck = []Check{
 		"Gauge",
 		"this is a test check a",
 		prometheus.NewGaugeVec(prometheus.GaugeOpts{}, []string{}),
-		[]map[string]string{map[string]string{"label1": "value1", "label2": "value2"}, map[string]string{"label1": "value3", "label2": "value4"}},
-		[]map[string]string{map[string]string{"label1": "value1", "label2": "value2"}},
+		[]map[string]string{{"label1": "value1", "label2": "value2"}, {"label1": "value3", "label2": "value4"}},
+		[]map[string]string{{"label1": "value1", "label2": "value2"}},
 		nil,
 		30,
 		0,
